convert automation/cookies into typescript
